Time: 0 ms (100.00%) | Memory: 8.5 MB (92.25%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
-    bool isPalindrome(long long x) {
+    bool isPalindrome(int x) {
         long long sum = 0;
-        long long temp = x;
+        int temp = x;
         while (x>0) {
             sum = sum * 10 + (x % 10);
             x /= 10;
